+++ src/main/java/es/hotel_back_v2/hotelV2/repositories/HabitacionRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface HabitacionRepository extends JpaRepository<Habitacion, Long> {
+public interface HabitacionRepository extends JpaRepository<Habitacion, Integer> {
     Optional<Habitacion> findByNumero(int numero);
 }
